feat: improve pseudocode parser
refactored the code to be class-based instead of function-based.

added the option to the file be updated or not.

moved the initial message and the final (success) message to the transpile function.

verbose_output and update_files variables are not been used because it simplifies the development of the code.